requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-chartjs==1.3
django-crispy-forms==1.7.2
idna==2.8
numpy==1.15.4
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
